move pkg people outside cmd. not sure how much I like this approach
